+++ demos/default/next.config.js
@@ -71,7 +71,7 @@ module.exports = {
   },
   // https://nextjs.org/docs/basic-features/image-optimization#domains
   images: {
-    domains: ['raw.githubusercontent.com', 'upload.wikimedia.org'],
+    domains: ['raw.githubusercontent.com'],
     remotePatterns: [
       {
         hostname: '*.imgur.com',