+++ demos/static-root/pages/index.js
@@ -1,5 +1,4 @@
 import Head from 'next/head'
-import Image from 'next/legacy/image'
 import { useRouter } from 'next/router'
 import styles from '../styles/Home.module.css'