+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [4.28.6](https://github.com/netlify/next-runtime/compare/plugin-nextjs-v4.28.5...plugin-nextjs-v4.28.6) (2022-11-07)
+
+
+### Bug Fixes
+
+* exclude swc by default, and don't exclude sharp if included ([#1745](https://github.com/netlify/next-runtime/issues/1745)) ([383c186](https://github.com/netlify/next-runtime/commit/383c186a988d972f84b1d04b1626f0670f71b065))
+
+## [4.28.5](https://github.com/netlify/next-runtime/compare/plugin-nextjs-v4.28.4...plugin-nextjs-v4.28.5) (2022-11-07)
+
+
+### Bug Fixes
+
+* resolve _app deps in API routes ([#1738](https://github.com/netlify/next-runtime/issues/1738)) ([df6fcda](https://github.com/netlify/next-runtime/commit/df6fcda60eab1b60c9cfb3710106e4bdec2ff1eb))
+
+## [4.28.4](https://github.com/netlify/next-runtime/compare/plugin-nextjs-v4.28.3...plugin-nextjs-v4.28.4) (2022-11-02)
+
+
+### Bug Fixes
+
+* only split extended routes to decrease build times ([#1731](https://github.com/netlify/next-runtime/issues/1731)) ([1e6fb8c](https://github.com/netlify/next-runtime/commit/1e6fb8cf2a1b00d57ae0a2df945514988ebb7dd3))
+
 ## [4.28.3](https://github.com/netlify/next-runtime/compare/plugin-nextjs-v4.28.2...plugin-nextjs-v4.28.3) (2022-10-27)